move mr_spark_harness.py out of mrjob/spark/ and into tests/
